ovp8xx/docker/requirements_deployment.txt
ifm3dpy==1.5.3
semver
colorama
typer
scp
pyyaml
pydantic